deleted extra files and added additional option for main class
