+++ specs/dont_use_recipient_spec.rb
@@ -1,0 +1,23 @@
+require_relative "test_helper"
+require "pry"
+# describe do
+#   # it "will raise an error when given an invalid channel" do
+#   #   error = Recipient.new(ES38768)
+
+#   #   VCR.use_cassette("slack-posts") do
+#   #     exception = expect {
+#   #       error.send_message("This post should not work", "invalid-channel")
+#   #     }.must_raise SlackApiError
+
+#   #     expect(exception.message).must_equal "Oops something went wrong"
+#   #   end
+#   # end
+
+#   # verify that it sends a message
+
+#   # what happens with invalid text
+
+#   # what happends with invalid token
+
+#   # look at required and option, and how program should respond
+# end
